requirements.txt
apiai==1.2.3
Flask==0.12.2
Flask-Script==2.0.5
requests==2.14.2
structlog==17.2.0
beautifulsoup4==4.6.0
gunicorn==19.7.1
python-dotenv==0.6.4
